// src/models/IoK8sApiAutoscalingV2HorizontalPodAutoscalerList/types.ts
import { IoK8sApiAutoscalingV2HorizontalPodAutoscalerModelOptions as IoK8sApiAutoscalingV2HorizontalPodAutoscaler } from '../IoK8sApiAutoscalingV2HorizontalPodAutoscaler';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiAutoscalingV2HorizontalPodAutoscalerListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiAutoscalingV2HorizontalPodAutoscaler>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}